Random route renamed to God route in backend; Created add form for Base, Campaign, Employee, Equipment entities; Render NumberInput and stub for SelectInput; Added downloading records in json

+++ course-04.09.20/frontend/src/components/EntityTable.jsx
@@ -1,9 +1,10 @@
 import React from "react"
-import {Table, message, Input, Button, Space, Layout, Popover, Checkbox} from "antd"
+import { Table, message, Input, Button, Space, Layout, Popover, Checkbox, InputNumber, Select, Tooltip } from "antd"
 import Highlighter from "react-highlight-words"
-import {DeleteOutlined, PlusOutlined, SearchOutlined} from "@ant-design/icons"
-import {Header, Content } from "antd/lib/layout/layout";
-import {EditableCell, EditableRow} from "./EditableRow.jsx";
+import { DeleteOutlined, PlusOutlined, SearchOutlined, DownloadOutlined } from "@ant-design/icons"
+import { Header, Content } from "antd/lib/layout/layout"
+import { EditableCell, EditableRow } from "./EditableRow.jsx"
+import { last, without } from "underscore"
 
 class EntityTable extends React.Component {
 
@@ -47,7 +48,7 @@ class EntityTable extends React.Component {
         onFilter: (value, record) =>
             record[dataIndex] ?
                 record[dataIndex].toString().toLowerCase().includes(value.toLowerCase())
-                : '',
+                : "",
         onFilterDropdownVisibleChange: visible => { if (visible) setTimeout(() => this.searchInput.select(), 100) },
         render: text =>
             this.state.searchedColumn === dataIndex ? (
@@ -55,7 +56,7 @@ class EntityTable extends React.Component {
                     highlightStyle={{ backgroundColor: '#ffc069', padding: 0 }}
                     searchWords={[this.state.searchText]}
                     autoEscape
-                    textToHighlight={text ? text.toString() : ''}
+                    textToHighlight={text ? text.toString() : ""}
                 />
             ) : (text)
     })
@@ -134,7 +135,9 @@ class EntityTable extends React.Component {
                 {
                     ...modifiedColumn,
                     filters: filters[strKey],
-                    onFilter: (v, r) => r[strKey] === v
+                    onFilter: (v, r) => r[strKey] === v,
+                    //TODO: нормальный select
+                    render: t => <Select defaultValue={t}/>
             } :
                 {
                     ...this.getColumnSearchProps(strKey),
@@ -144,6 +147,11 @@ class EntityTable extends React.Component {
                 modifiedColumn = {
                     ...modifiedColumn,
                     render: t => <Checkbox checked={t}/>
+                }
+            if (typeof object[key] === "number" && modifiedColumn.editable)
+                modifiedColumn = {
+                    ...modifiedColumn,
+                    render: t => <InputNumber defaultValue={t}/>
                 }
             if (modifiedColumn.editable)
                 modifiedColumn = {
@@ -168,7 +176,9 @@ class EntityTable extends React.Component {
         и вовсе это запрещает. Т.к. кардинально менять api бэка не хотелось, решил костыльно передавать json-массив как
         параметр запроса. */
         const url = this.props.presenter.url + "?=" + new URLSearchParams({
-            "ids": '{ "selectedIds": [] }'
+            "ids":  JSON.stringify({
+                selectedIds: []
+            })
         })
         const req = { method: "GET", mode: "cors" }
         fetch(url, req).then(res => res.json()).then(
@@ -194,27 +204,56 @@ class EntityTable extends React.Component {
     }
 
     removeRecords = () => {
-        const url = this.props.presenter.url
-        const req = {
-            method: "DELETE",
-            mode: "cors",
-            headers: { "Content-Type": "application/json" },
-            body: JSON.stringify({ droppedIds: this.state.selectedRowKeys })
+        if (this.state.selectedRowKeys.length === 0)
+            message.success({ top: 100, content: "Can't delete, nothing selected", marginTop: "20vh" })
+        else {
+            const url = this.props.presenter.url
+            const req = {
+                method: "DELETE",
+                mode: "cors",
+                headers: { "Content-Type": "application/json" },
+                body: JSON.stringify({ droppedIds: this.state.selectedRowKeys })
+            }
+            fetch(url, req).then(res => res.text()).then(
+                data => {
+                    message.success({ top: 100, content: data, marginTop: "20vh" })
+                    this.setState(
+                        { items: this.state.items.filter(x =>
+                                !this.state.selectedRowKeys.includes(x[this.props.presenter.idField]))
+                    })
+                },
+                error => {
+                    this.setState({ error: error })
+                    message.error({ top: 24, content: error.message.toLocaleString(), marginTop: "20vh" })
+                })
         }
-        fetch(url, req).then(res => res.text()).then(
-            data => {
-                message.success({ top: 100, content: data, marginTop: "20vh" })
-                const newItems = this.state.items.filter(x => !this.state.selectedRowKeys.includes(x[this.props.presenter.idField]))
-                this.setState({ items: newItems })
-            },
-            error => {
-                this.setState({ error: error })
-                message.error({ top: 24, content: error.message.toLocaleString(), marginTop: "20vh" })
-            })
     }
 
     addRecord = () => {
         //TODO: взять колонки, убрать ту, где id собрать элемент
+    }
+
+    downloadRecords = () => {
+        const url = this.props.presenter.url + "?=" + new URLSearchParams({
+            "ids": JSON.stringify({
+                selectedIds: (this.state.selectedRowKeys === 0) ? [] : this.state.selectedRowKeys
+            })
+        })
+        const req = { method: "GET", mode: "cors" }
+        fetch(url, req).then(res => res.text()).then(
+            data => {
+                require("file-saver").saveAs(
+                    new Blob([data], { type: "text/json;charset=utf-8" }),
+                    last(this.props.presenter.url.split("/")) + ".json"
+                )
+            },
+            error => {
+                message.error({
+                    top: 24,
+                    content: error.message.toLocaleString(),
+                    marginTop: "20vh"
+                })
+            })
     }
 
     // Вызывается лишь раз, при создании компонента
@@ -234,8 +273,9 @@ class EntityTable extends React.Component {
         return <Layout className="site-layout">
             <Header style={{ position: 'fixed', zIndex: 1, width: '100%' }}>
                 <Space size={"middle"}>
-                    <Popover trigger="click" content={<EntityCreator/>}>
+                    <Popover trigger="click" content={this.props.presenter.creator}>
                         <Button
+                            type="primary"
                             icon={<PlusOutlined/>}
                             ghost={true}
                             onClick={this.addRecord}>
@@ -249,6 +289,14 @@ class EntityTable extends React.Component {
                         onClick={this.removeRecords}>
                         Remove
                     </Button>
+                    <Tooltip title="All objects from table will be downloaded if nothing is selected">
+                        <Button
+                            icon={<DownloadOutlined/>}
+                            ghost={true}
+                            onClick={this.downloadRecords}>
+                            Download as json
+                        </Button>
+                    </Tooltip>
                 </Space>
             </Header>
             <Content style={{ marginTop: "64px" }}>
@@ -265,17 +313,6 @@ class EntityTable extends React.Component {
                 </div>
             </Content>
         </Layout>
-    }
-}
-
-class EntityCreator extends React.Component {
-    /*Это пример плохого дизайна 2, но уже со стороны фронта*/
-    render() {
-        return <div>
-            <Table
-            />
-            <Button>Create</Button>
-        </div>
     }
 }
